Update the ITCH accuracy statistics analysis
The statistics were made to compare my results with the results of
S. Wang in her paper 'Cross-response in correlated financial markets:
individual stocks'. We used the definitions of equation (1) , (2) and
(3) to achieve the results. The analysis were made using the ITCH data,
as it has the real trades with their corresponding prices, volumes and
times. So it is possible to test the model and compare the results to
know the accuracy of the classification