Update the information and implementation of the TAQ data load
The TAQ data is stored in a compressed way, and it is needed a `C++` to decompress
the data. For my analysis I only decompressed the AAPL and MSFT tickers in `CSV`
format to work with them in Python.

The TAQ data have two files, one with the trades information and another
with the quoutes information. Each quotes file is really big (> 5 GB)
as it cointains the information of a full year and the trades files are
around ~ 1 GB, in our case we analyze the year 2008.

I try to load the data in different ways to avoid the memory of the PC
overloads with this files, and I found a module in `Python` called
`Dask` that is a flexible library for parallel computing in Python.

However, the results were not consistent, so I had to load the data and
immediatly convert the interest columns in the dask dataframe to pandas
to avoid the errors I was having. I would be interesting to analyze
where was the exact point where the results just broke up, to obtain the
best performance of dask (future task (?)).

Then I just save the columns of the pandas dataframe with some filters
in numpy arrays to save the relevant information I would use in further
analysis.

The final part of the code return a list with the days in each day that
have data. Normally, on weekends the market is close and strangely
there are some days in the week that there is no data.