Implement a test to confirm the number of identified trades
Test to find the number of trades in an ITCH data file, assuming that
there is a trade when the type of the order is "E" -- Execute
outstanding order in part or "F" -- Execute outstanding order in full.

I used the stocks

- AAPL (07.01.2008)
- AAPL (02.06.2008)
- GS (07.10.2008)
- GS (10.12.2008)
- XOM (11.02.2008)
- XOM (04.08.2008)

#### Conclusion

The number of identified trades is equal to the identified trades
reported by S. Wang in her paper.