Implement a new algorithm to identify the trade signs
I implement a new approach to find all the trade signs in the market
time as the  previous implementation was having problems identifying the
trades and the corresponding trade signs.

I load the dataset as a pandas dataframe and divide it in two
dataframes, one for the limit orders and another for the trade orders.
As the limit order is a huge amount of data, I reduced to only take care
of the data related with the trade orders.

I identified in each step of the `for loop` the limit order associated
with the corresponding trade order using the order id. With this I save
the time, trade sign, volume and price.

With the volumes there is a special case if there is a full or partial
transaction. So it is necessary to organize the data after this happens.

When a full transaction is made, the order is change to zero to avoid
repeating the value.