Try to improve the time in the identification of trades
As the pandas implementation is really slow, I implement a version to
identify the trades using only numpy. However, just loading the data
the time it takes is greater than Pandas. With this in mind I just stop
the implementation, and the new idea is to combine pandas with
numpy to see if I can get an improvement