Implement the trade classification using pandas and numpy
In this implementation I mixed pandas and numpy. I loaded and filtered
the data using pandas because is fast and easy to divide the data in
what I needed. Then I convert each column of the pandas data in a
numpy array. With numpy array the searching of the order values
number is fast and the `for` loop run fast in comparison with the
implementation using pandas only.

This implementation will be used to test the accuracy of the model used by S. Wang.