Review TAQ self response data
The previous version was adding zeros in the calculation that were not
necessary.

I analyzed the extreme cases in the log returns calculations but none of
them were a problem. The analysis start with the fact that the midpoint
price is never zero, so the division in the logarithm is always defined.

1. If the midpoint price with the time lag is smaller (<<) than the
midpoint price, the division is zero and the log is -Inf.

2. If the midpoint price with the time lag is greater (>>) than the
midpoint price, the division will diverge, and the log diverges.

3. If the midpoint price with the time lag is equal to the midpoint
price, the division is one and the log is 0. Then it does not add
nothing to the calculation.

In cases 1 and 2 the probability that this happens is really small, as the
changes in the midpoint price stay in the same scale.

To solve the problem of the appended zeros, my strategy is to take the
original array of trade signs with all its zeros and then reduce it
length by the corresponding tau value and multiply it by the calculated
log return without append zeros.

Finally, the resulting array values are added and divided by the total
amount of reduced trades signs different to zero.

With this approximation, when tau is small there will not be a huge
impact, and with the last value of tau (~1000) maybe there is a small
problem as the maximum total of time is 22200 seconds and the trades
are of the order of ~14000 (the maximum tau is around the 7% of
identified trades). But in practice it will no add a huge error.

However, with this changes the results are not different and are still
far away from S. Wang results