Review TAQ cross response data
The results are quite good. The behavior of the two plots are "similar"
to the results of S. Wang. The shape can be seen but the intensity of
the response is slower.

This can be justified by the way I calculate the response or by the
differences in the trade signs identification.

Further analysis of the S. Wang algorithm is requiered