Update TAQ trade classification
Again I reviewed the code trying to find a bug in the TAQ trade sign
classification. However I end with the same results for the first stage
making the classification for sucesive trades.

In the second part I implement the classification as S. Wang did it her
paper. This give me some variations in the results but not to
considerable. I updated the names of the functions and delete the
function thta transforms from millisecond to seconds, as it was not
longer required.