Create a jpnb to compare classification results
Due to the differences in the results of the accuracy of trade sign
classification, I create a jpnb to compare value to value where were the
diffences and find out the reasons of them.

The first step was to obtain the results in the same format of S. Wang.
I runned my algorithm and used pandas dataframes. Then I loaded S. Wang
results also in pandas dataframes, and then I started to compare every
single value of both results.

For the times and the reference trade signs values used for transactions
and persecond, all the values where the same (similarity of 100%).
However for the calculate values using the equations of the paper, the
values were not the same. Then the problem was in the implementation of
the equations.

The values that had different numbers have an interesting characteristic
The predicted value I used is accorded to the theory but gives a bad
classification. Incredibly, the S. Wang algorithm predict this values
without logic.

I still do not have the answer but this check is a start point to
understand the ground of the differences.