Add the hidden orders to compute the accuracy
Analyzing the code of S. Wang, there is a little detail when the trades
signs will be computed for the first time. Despite we do not use hidden
orders in our calculations, the ITCH gives the time, price and volume of
them, and it is completely logic that the hidden orders have an impact
the prices of the order book, as the consume stocks from them. However I
was thinking that as the article never mentioned the hidden orders, and
in fact there is a paragraph that says the hidden orders are not used,
then probably the hidden orders were not useful. This is the main reason
I was obtaining wrong values compared with the analysis of S. Wang.

Now the millisecond results are exactly the same, and the second results
are similar. I will test the second results in the future, but at least the
first part is working.

I updated the corresponding files with the new algorithm.