Find the bug in S. Wang code for trade classification
Knowing that the millisecond results were the same, I started to analyze
the S. Wang code. The line to compute the second results was not
complicated to understand, and in theory it was the same implementation
I had in my code. However the results were different. As I was reviewed
my code thousand times, I start to doubt of the S. Wang code.

Finally, I found that the assigned variables to calculate the eq. 2 and
3 were wrong. S. Wang confused the theoric and experimental values. I
corrected S. Wang code and then again compare the results. For eq. 2 I
have 100% similarities and with eq. 3 I have 91% similarities. The
general accuracy of the classification improved by ~1%.

Know I can say my code can accurately classify the trade signs.
The next step is test the cross-responses.