Find the bug in the TAQ data load
Comparing the results of S. Wang with my results in detail, the data of
the time was the same, but the values of the prices differ. In a close
revision, the values were the same, but they did not have the same order.

There was a mistake in my code or in S. Wang code. The load function of
S. Wang loaded the data in order, and from the beggining I doubted about
the task module. I implemented an alternative data load to test which of
the both results was the wrong one.

The test showed that S. Wang results were correct and that I had a
problem with my load method. I change the code to the new load method
to have the same basic data, so I can reproduce her final results.