Change the TAQ data extraction implementation
Considering the awful behavior of dask, I implement an extraction code
reading the data line per line and making in the reading the complete
filtering of the times and dates to obtain the data for each date.

With this implementation there is no problem to use the python parallel
module to make the extraction of the data faster.