Update the TAQ modules to check the corrected data
After finding the bug of the load data, I used again the TAQ module.
However, it was a little bit broken so I fixed some crucial points to
test the new data.

The signal is much better than the previous one, however is not yet the
same. The result is good but a further revision is required.