Update taq wang comparison
Finally my trade signs and midpoint prices are the same with S. Wang
results, so the cross response function is almost the same. The problem
for the trade signs was that my load method was
wrong, and for the cross response function that the time was shifted
between trade signs and midpoint prices. After this corrections, the
cross response function is almost the same, not equal probably for the
small difference in the averaging number, but at the end not
representative