Finish ITCH trade sign classification test
I reviewed the implementation of the ITCH trade sign classification,
found a mistake in my interpretation and corrected the results.

The results between my implementation and the paper are still not the
same, but using the correction of the paper implementation I have
100% accuracy.

I made a jupyter notebook to show the difference in the results and to
explain why I change the implementation of the paper's code.